Pass HIP_PLATFORM to bazel build
This PR is to address the TF build failure reported in JIRA ticket:
http://ontrack-internal.amd.com/browse/SWDEV-241967

The valid set of values for the env var HIP_PLATFORM is being updated, and as a consequence `hcc` will no longer a valid value (staring with ROCm 3.7).
For ROCm 3.7, we will need to change the configure.py script such that the env var HIP_PLATFORM=amd is explicitly when invoked with TF_NEED_ROCM=1.

In the meantime (i.e. until we switch to ROCm 3.7), we need to provide a way for passing HIP_PLATFORM to bazel (to help QA in testing ROCm 3.7). This change provides that way.
If the HIP_PLATFORM env var is set in the shell that invokes the "configure.py" script, it will pass it through to bazel.