Support iterator checkpointing with large caches.
Previously there was a 2GB limit on cache size, due to protobuf serialization limits. Now we write each element to a new protobuf, so that we can checkpoint the cache as long as individual elements are less than 2GB.

The utilities added to dataset_utils can be used to simplify checkpointing in many ops that store buffers of elements, such as prefetch, parallel interleave, parallel map, and shuffle. I will submit additional CLs to update those datasets.

PiperOrigin-RevId: 321021511
Change-Id: Ib8cbb0979e8b1fa9b37b82837fc37b4c85f998c2