Fixes a misleading error message when creating a functional `tf.keras.Model()`.
The Keras functional API uses the keyword arguments `inputs` and `outputs`. Currently however, if someone misspells `outputs` (e.g. as `output`), they get the following error: `TypeError: ('Keyword argument not understood:', 'inputs')`. This is confusing as it suggests that there is a problem with `inputs`, not `output`.

This error might ideally be surfaced where it is detected (and then ignored in a `try` statement) in `Functional.__init__()`, however that would require a larger change.

Instead this error message is fixed by telling `Model.__init__()` to validate `"inputs'` and `'outputs'` with its `kwargs`. This is less ideal because these arguments are not supposed to be passed to `Model.__init__()`, but as the user thinks that they are calling simply calling `Model.__init__()` it should not cause them any confusion.

This fixes Keras [issue 13743](https://github.com/keras-team/keras/issues/13743).

PiperOrigin-RevId: 321668253
Change-Id: Ideff3cd9298f573b633a2e6e821fa77b1c862570