Add a `convert_to_tensor` to the start of Tensor.__getitem__ (_slice_helper) to make sure it dispatches directly, rather than letting the nested tf.strided_slice trigger dispatching.
This is important because `tensor.__getitem__` does some input arg manipulation before getting to the `tf.strided_slice`. So, when we try to run the traced code using the args provided to `strided_slice` (e.g. for KerasTensors), we lose information about constants that TPUs need to compile graphs involving shape manipulation. Tracing `__getitem__` and its input args directly does not seem to run into this problem.

(Note: this TPU situation is separate from the shape value inferring we do in KerasTensors during Functional API construction/tracing time. This happens at model run-time when running the already-traced code)

To get this all to work correctly in practice, this CL also has to:

* Add tf.nest support for flattening & packing python `slice` objects, in case the slice object contains symbolic tensors/values to trace

* Add serialization/deserialization support for `slice` and `ellipsis` objects in Keras

PiperOrigin-RevId: 322239438
Change-Id: If9b72368dff8bd50b61a1adbc6162f0a8da684d3