Add a StartAbort() method to CollectiveParamResolver with local implementation
This is part of changes towards enabling aborting collectives when the group is unhealthy. BaseCollectiveExecutor used to only abort the BufRendezvous, but the collectives may be waiting on param resolution as well.

This change only contains implementation for CollectiveParamResolverLocal, not CollectiveParamResolverDistributed. This means we don't abort the RPCs to the collective leader, nor the RPCs to get device attributes of members of the group. For the former the abortion is delayed until the collective leader is offline (those RPCs are fail_fast), for the latter the abortion is delayed until it gets all device attributes (those RPCs are not fail fast). These are not ideal but should be fine in a deployment where workers will eventually come back.

I'll implement CollectiveParamResolverDistributed::StartAbort() in a later change.

PiperOrigin-RevId: 322455670
Change-Id: I494d6b367e90a7a14a10dd20e111b37e892e45b9