[XLA] Fix a spurious verification failure with nested conditionals.
When allocating alternate memory for conditionals, we assume the conditional
will always evict the value back to default memory before the end of the called
computations. In other words, the output of conditional can never get alternate
memory allocations due to difficulties with aliasing, but the inputs can. When
verifying the correctness of memory space assignment, we have to split the
conditional uses by the called computations and find the time bounds in each
computation that the buffer in alternate memory is used. We previously didn't do
this splitting for nested conditionals, causing b/161935244. Now we do this
splitting recursively.

PiperOrigin-RevId: 322865495
Change-Id: I522ccd9e54e73ee9d3b53997fb28198651338909