Share ownership of CollectiveImplementation in `BaseCollectiveExecutor::ExecuteAsync`.
Without this change, we would sometimes get a segfault for a short-running
collective.  The reason is that the CUDA kernel can complete and trigger the
done callback, which deletes `col_impl` object, before the TF kernel has
finished execution.

This change also updates collective_ops_gpu_test to use EagerContext for
checking number of available GPUs.

Resolves #41113.

PiperOrigin-RevId: 323506550
Change-Id: I414f2b79875f51963e7d5e2b454e106d88b511be