Support TF NumPy arrays in tf.autodiff.ForwardAccumulator
Expands composites in its constructor (like GradientTape.watch does) and special-cases them in ForwardAccumulator.jvp for fetching a tangent. Expanding in .jvp would mean returning SparseTensor(indices=None) which seems a bit weird.

This fixes a tangential issue with unconnected_gradients=zero and variables, so I've included a test for that.

PiperOrigin-RevId: 324008908
Change-Id: I0f4b22146519cfa38235d4e6440c12f8220c8a18