[tf.data service] Update dispatcher_impl to use dispatcher_state.
This CL updates dispatcher_impl to store its dataset and job state in `DispatcherState`, and only update the state by applying state updates.

The CL also changes the Task and Worker classes to be simple structs, to be consistent with the Dataset and Job classes in `DispatcherState`.

No external-facing functionality is changed. The refactored functionality is covered by the tests in data_service_ops_test.py

PiperOrigin-RevId: 324111209
Change-Id: I6163565c10e074d5973207039ff65ba0be071d21