Fix issue in tf.string.format wheree unicode is not rendered correctly
This PR tries to address the issue raised in 42001 where
unicode tensor is not rendered correctly in tf.string.format.

The issue was that tf.string.format incorrectly escape the tensor.

This PR address the issue by adding escape option in tensor print
function to selectively control the escape.

This PR fixes 42001.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>